refactor: user -> users
