+++ src/main/java/capstone/examlab/users/argumentresolver/Login.java
@@ -1,4 +1,4 @@
-package capstone.examlab.user.argumentresolver;
+package capstone.examlab.users.argumentresolver;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
