+++ src/main/java/capstone/examlab/users/domain/User.java
@@ -1,4 +1,4 @@
-package capstone.examlab.user.domain;
+package capstone.examlab.users.domain;
 
 import jakarta.persistence.MappedSuperclass;
 import lombok.Data;
