+++ src/main/java/capstone/examlab/users/dto/LoginDto.java
@@ -1,22 +1,16 @@
-package capstone.examlab.user.dto;
+package capstone.examlab.users.dto;
 
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
 import lombok.Data;
 
 @Data
-public class UserAddDto {
+public class LoginDto {
 
     @NotBlank
     @Email
     private String userId;
 
     @NotBlank
-    private String name;
-
-    @NotBlank
     private String password;
-
-    @NotBlank
-    private String passwordConfirm;
 }
