+++ src/test/java/capstone/examlab/exams/controller/ExamsControllerTest.java
@@ -2,7 +2,7 @@ package capstone.examlab.exams.controller;
 
 import capstone.examlab.exams.repository.ExamRepository;
 import capstone.examlab.exams.service.ExamsService;
-import capstone.examlab.user.argumentresolver.LoginUserArgumentResolver;
+import capstone.examlab.users.argumentresolver.LoginUserArgumentResolver;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Tag;
